[refactor] move Chain, Config type to packages/core/src/constants/types for outside access
