+++ packages/core/src/constants/index.ts
@@ -1,4 +1,5 @@
 export * from './abi'
-export * from './Chain'
+export * from './type/Chain'
+export * from './type/Config'
 export * from './chainId'
 export * from './currencies'
