+++ packages/core/src/model/chain/bsc.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const BSC: Chain = {
   chainId: 56,
