+++ packages/core/src/model/chain/ethereum.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Mainnet: Chain = {
   chainId: 1,
