+++ packages/core/src/model/chain/fantom.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Fantom: Chain = {
   chainId: 250,
