+++ packages/core/src/model/chain/local.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Localhost: Chain = {
   chainId: 1337,
