+++ packages/core/src/model/chain/polygon.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Polygon: Chain = {
   chainId: 137,
