+++ packages/core/src/model/chain/songbird.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Songbird: Chain = {
   chainId: 19,
