+++ packages/core/src/model/chain/theta.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const Theta: Chain = {
   chainId: 361,
