+++ packages/core/src/model/chain/xdai.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../constants/Chain'
+import { Chain } from '../../constants'
 
 export const xDai: Chain = {
   chainId: 100,
