+++ packages/core/src/model/config/default.ts
@@ -1,6 +1,4 @@
-import { DEFAULT_SUPPORTED_CHAINS } from '../../constants/chainId'
-
-import { FullConfig } from './Config'
+import { DEFAULT_SUPPORTED_CHAINS, FullConfig } from '../../constants'
 
 export const DEFAULT_CONFIG: FullConfig = {
   pollingInterval: 15000,
