+++ packages/core/src/model/index.ts
@@ -1,5 +1,4 @@
 export * from './Currency'
 export * from './CurrencyValue'
-export * from './config/Config'
 export * from './TransactionStatus'
 export * from './TransactionOptions'
